backend/requirements.txt
azure-functions
azure-functions-durable
azure-storage-blob
azure-ai-formrecognizer
azure-ai-contentsafety
azure-identity
openai
filetype
Pillow
pypdf
pandas
requests